+++ rest_framework_services/__init__.py
@@ -67,4 +67,4 @@ __all__ = [
     "update_from_input",
 ]
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"